move: seperate kebob btn, modal components in qna detail
